requirements.txt
streamlit
google-genai
chromadb
sentence-transformers
